+++ core/jobs.h
@@ -1,0 +1,14 @@
+//* jobs
+//      ...
+
+#include "fresa_types.h"
+
+namespace fresa::jobs
+{
+    //- job queue
+
+    //- fiber pool
+
+    //- worker threads
+
+}